+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added imagePickerBuilder to the builders in the ChatOptions to override the image picker with a custom implementation that needs to return a Future<Uint8List?>
 - Changed the ChatBottomInputSection to be multiline and go from 45px to 120px in height depending on how many lines are in the textfield
 - Added chatScreenBuilder to the userstory configuration to customize the specific chat screen with a ChatModel as argument
+- Added senderTitleResolver to the ChatOptions to resolve the title of the sender in the chat message
 
 ## 4.0.0
 - Move to the new user story architecture